zNPCCommon definition removed, now included from zNPCTypeCommon. Moved en_SM_NOTICES to zNPCTypeCommon to avoid recursive include.
