+++ src/SB/Game/zNPCTypeCommon.h
@@ -10,7 +10,6 @@
 #include "xSFX.h"
 
 #include "zNPCSndTable.h"
-#include "zNPCSpawner.h"
 #include "zMovePoint.h"
 #include "zShrapnel.h"
 
@@ -338,6 +337,20 @@ enum en_NPC_MSG_DATA
     NPC_MDAT_AREANOTIFY,
     NPC_MDAT_NOMORE,
     NPC_MDAT_FORCE = 0x7fffffff
+};
+
+enum en_SM_NOTICES
+{
+    SM_NOTE_NPCDIED,
+    SM_NOTE_NPCSTANDBY,
+    SM_NOTE_NPCALIVE,
+    SM_NOTE_DUPPAUSE,
+    SM_NOTE_DUPRESUME,
+    SM_NOTE_DUPSETDELAY,
+    SM_NOTE_DUPDEAD,
+    SM_NOTE_KILLKIDS,
+    SM_NOTE_NOMORE,
+    SM_NOTE_FORCE = 0x7fffffff
 };
 
 struct zNPCLassoInfo
